# Move the execute function to use the new interface