--- a/src/infer_request.cc
+++ b/src/infer_request.cc
@@ -29,7 +29,6 @@
 
 #include "pb_utils.h"
 #ifdef TRITON_PB_STUB
-#include "infer_response.h"
 #include "pb_stub.h"
 #endif
 
@@ -96,204 +95,155 @@ InferRequest::SetFlags(uint32_t flags)
 }
 
 void
-InferRequest::SaveToSharedMemory(
-    std::unique_ptr<SharedMemory>& shm_pool, Request* request_shm)
+InferRequest::SaveToSharedMemory(std::unique_ptr<SharedMemoryManager>& shm_pool)
 {
-  request_shm->correlation_id = this->CorrelationId();
-  off_t id_offset;
-  SaveStringToSharedMemory(shm_pool, id_offset, this->RequestId().c_str());
-  request_shm->id = id_offset;
-  request_shm->requested_output_count = this->RequestedOutputNames().size();
-  off_t requested_output_names_offset;
-  off_t* requested_output_names;
-  shm_pool->Map(
-      (char**)&requested_output_names,
-      sizeof(off_t) * request_shm->requested_output_count,
-      requested_output_names_offset);
-  request_shm->flags = Flags();
+  AllocatedSharedMemory<InferRequestShm> infer_request_shm =
+      shm_pool->Construct<InferRequestShm>();
 
-  request_shm->requested_output_names = requested_output_names_offset;
+  infer_request_shm.data_->correlation_id = CorrelationId();
+
+  std::unique_ptr<PbString> request_id_shm =
+      PbString::Create(shm_pool, RequestId());
+
+  infer_request_shm.data_->id = request_id_shm->ShmOffset();
+  infer_request_shm.data_->requested_output_count =
+      RequestedOutputNames().size();
+
+  AllocatedSharedMemory<bi::managed_external_buffer::handle_t>
+      output_names_handle =
+          shm_pool->ConstructMany<bi::managed_external_buffer::handle_t>(
+              RequestedOutputNames().size());
+  infer_request_shm.data_->requested_output_names = output_names_handle.handle_;
+
   size_t i = 0;
+  std::vector<std::unique_ptr<PbString>> requested_output_names_shm(
+      RequestedOutputNames().size());
   for (auto& requested_output_name : requested_output_names_) {
-    SaveStringToSharedMemory(
-        shm_pool, requested_output_names[i], requested_output_name.c_str());
+    std::unique_ptr<PbString> requested_output_name_shm =
+        PbString::Create(shm_pool, requested_output_name);
+    (output_names_handle.data_.get())[i] =
+        requested_output_name_shm->ShmOffset();
+    requested_output_names_shm.emplace_back(
+        std::move(requested_output_name_shm));
     i++;
   }
 
-  request_shm->requested_input_count = this->Inputs().size();
-  request_shm->model_version = this->model_version_;
-  SaveStringToSharedMemory(
-      shm_pool, request_shm->model_name, this->model_name_.c_str());
+  infer_request_shm.data_->input_count = Inputs().size();
+  infer_request_shm.data_->model_version = model_version_;
+
+  std::unique_ptr<PbString> model_name_shm =
+      PbString::Create(shm_pool, ModelName());
+  infer_request_shm.data_->model_name = model_name_shm->ShmOffset();
+
+  AllocatedSharedMemory<bi::managed_external_buffer::handle_t>
+      input_tensors_handle =
+          shm_pool->ConstructMany<bi::managed_external_buffer::handle_t>(
+              Inputs().size());
+
+  infer_request_shm.data_->inputs = input_tensors_handle.handle_;
+  i = 0;
+  for (auto& input : Inputs()) {
+    input->SaveToSharedMemory(shm_pool);
+    (input_tensors_handle.data_.get())[i] = input->ShmOffset();
+    i++;
+  }
+
+
+  // Save the references to shared memory.
+  infer_request_shm_ = std::move(infer_request_shm);
+  infer_request_shm_ptr_ = infer_request_shm.data_.get();
+  request_id_shm_ = std::move(request_id_shm);
+  output_names_handle_shm_ = std::move(output_names_handle);
+  requested_output_names_shm_ = std::move(requested_output_names_shm);
+  input_tensors_handle_ = std::move(input_tensors_handle);
+  input_tensors_handle_ptr_ = input_tensors_handle.data_.get();
 }
 
 std::unique_ptr<InferRequest>
 InferRequest::LoadFromSharedMemory(
-    std::unique_ptr<SharedMemory>& shm_pool, off_t request_offset,
-    std::shared_ptr<std::mutex>& cuda_ipc_open_mutex,
-    std::shared_ptr<std::mutex>& cuda_ipc_close_mutex)
+    std::unique_ptr<SharedMemoryManager>& shm_pool,
+    bi::managed_external_buffer::handle_t request_offset)
 {
-  Request* request;
-  shm_pool->MapOffset((char**)&request, request_offset);
+  AllocatedSharedMemory<InferRequestShm> infer_request_shm =
+      shm_pool->Load<InferRequestShm>(request_offset);
+  std::unique_ptr<PbString> request_id_shm =
+      PbString::LoadFromSharedMemory(shm_pool, infer_request_shm.data_->id);
 
-  char* id = nullptr;
-  LoadStringFromSharedMemory(shm_pool, request->id, id);
+  AllocatedSharedMemory<bi::managed_external_buffer::handle_t>
+      input_tensors_handle =
+          shm_pool->Load<bi::managed_external_buffer::handle_t>(
+              infer_request_shm.data_->inputs);
 
-  uint32_t requested_input_count = request->requested_input_count;
+  AllocatedSharedMemory<bi::managed_external_buffer::handle_t>
+      output_names_handle =
+          shm_pool->Load<bi::managed_external_buffer::handle_t>(
+              infer_request_shm.data_->requested_output_names);
 
-  std::vector<std::shared_ptr<PbTensor>> py_input_tensors;
-  for (size_t input_idx = 0; input_idx < requested_input_count; ++input_idx) {
-    std::shared_ptr<PbTensor> pb_input_tensor = PbTensor::LoadFromSharedMemory(
-        shm_pool, request->inputs + sizeof(Tensor) * input_idx,
-        cuda_ipc_open_mutex, cuda_ipc_close_mutex);
-    py_input_tensors.emplace_back(std::move(pb_input_tensor));
-  }
-
-  std::vector<std::string> requested_output_names;
-  uint32_t requested_output_count = request->requested_output_count;
-  off_t* output_names;
-  shm_pool->MapOffset((char**)&output_names, request->requested_output_names);
+  std::vector<std::unique_ptr<PbString>> requested_output_names_shm;
+  uint32_t requested_output_count =
+      infer_request_shm.data_->requested_output_count;
 
   for (size_t output_idx = 0; output_idx < requested_output_count;
        ++output_idx) {
-    char* output_name = nullptr;
-    LoadStringFromSharedMemory(shm_pool, output_names[output_idx], output_name);
-    requested_output_names.emplace_back(output_name);
+    std::unique_ptr<PbString> pb_string = PbString::LoadFromSharedMemory(
+        shm_pool, (output_names_handle.data_.get())[output_idx]);
+    requested_output_names_shm.emplace_back(std::move(pb_string));
   }
 
-  char* model_name;
-  LoadStringFromSharedMemory(shm_pool, request->model_name, model_name);
-  return std::make_unique<InferRequest>(
-      id, request->correlation_id, std::move(py_input_tensors),
-      requested_output_names, model_name, request->model_version,
-      request->flags);
+  std::vector<std::shared_ptr<PbTensor>> input_tensors;
+  for (size_t input_idx = 0; input_idx < infer_request_shm.data_->input_count;
+       ++input_idx) {
+    std::shared_ptr<PbTensor> input_tensor = PbTensor::LoadFromSharedMemory(
+        shm_pool, (input_tensors_handle.data_.get())[input_idx]);
+    input_tensors.emplace_back(std::move(input_tensor));
+  }
+
+  std::unique_ptr<PbString> model_name_shm = PbString::LoadFromSharedMemory(
+      shm_pool, infer_request_shm.data_->model_name);
+
+  return std::unique_ptr<InferRequest>(new InferRequest(
+      infer_request_shm, request_id_shm, requested_output_names_shm,
+      model_name_shm, output_names_handle, input_tensors_handle,
+      input_tensors));
 }
 
-#ifdef TRITON_PB_STUB
-std::unique_ptr<InferResponse>
-InferRequest::Exec()
+InferRequest::InferRequest(
+    AllocatedSharedMemory<InferRequestShm>& infer_request_shm,
+    std::unique_ptr<PbString>& request_id_shm,
+    std::vector<std::unique_ptr<PbString>>& requested_output_names_shm,
+    std::unique_ptr<PbString>& model_name_shm,
+    AllocatedSharedMemory<bi::managed_external_buffer::handle_t>&
+        output_names_handle_shm,
+    AllocatedSharedMemory<bi::managed_external_buffer::handle_t>&
+        input_tensors_handle,
+    std::vector<std::shared_ptr<PbTensor>>& input_tensors)
+    : infer_request_shm_(std::move(infer_request_shm)),
+      request_id_shm_(std::move(request_id_shm)),
+      requested_output_names_shm_(std::move(requested_output_names_shm)),
+      model_name_shm_(std::move(model_name_shm)),
+      output_names_handle_shm_(std::move(output_names_handle_shm)),
+      input_tensors_handle_(std::move(input_tensors_handle))
+
 {
-  ResponseBatch* response_batch = nullptr;
-  bool responses_is_set = false;
-  std::unique_ptr<Stub>& stub = Stub::GetOrCreateInstance();
-  std::unique_ptr<SharedMemory>& shm_pool = stub->GetSharedMemory();
+  infer_request_shm_ptr_ = infer_request_shm_.data_.get();
+  output_names_handle_shm_ptr_ = output_names_handle_shm_.data_.get();
+  input_tensors_handle_ptr_ = input_tensors_handle_.data_.get();
+  inputs_ = std::move(input_tensors);
 
-  try {
-    py::gil_scoped_release release;
-    std::unique_ptr<IPCMessage> ipc_message =
-        std::make_unique<IPCMessage>(shm_pool, true /* inline_response */);
-    bool has_exception = false;
-    PythonBackendException pb_exception(std::string{});
-
-    ipc_message->Command() =
-        PYTHONSTUB_CommandType::PYTHONSTUB_InferExecRequest;
-
-    RequestBatch* request_batch;
-    shm_pool->Map(
-        (char**)&request_batch, sizeof(RequestBatch), ipc_message->Args());
-    request_batch->batch_size = 1;
-
-    Request* request;
-    shm_pool->Map((char**)&request, sizeof(Request), request_batch->requests);
-
-    request->requested_input_count = this->Inputs().size();
-    Tensor* tensors;
-    bool has_gpu_tensor = false;
-    shm_pool->Map(
-        (char**)&tensors, sizeof(Tensor) * request->requested_input_count,
-        request->inputs);
-
-    size_t i = 0;
-    for (auto& input_tensor : inputs_) {
-      input_tensor->SaveToSharedMemory(
-          shm_pool, &tensors[i], true /* copy_cpu */, false /* copy_gpu */);
-      if (!input_tensor->IsCPU()) {
-        has_gpu_tensor = true;
-      }
-      ++i;
-    }
-
-    SaveToSharedMemory(shm_pool, request);
-    {
-      bi::scoped_lock<bi::interprocess_mutex> lock{
-          *(ipc_message->ResponseMutex())};
-      stub->SendIPCMessage(ipc_message);
-      ipc_message->ResponseCondition()->wait(lock);
-    }
-
-    if (has_gpu_tensor) {
-      try {
-        for (auto& input_tensor : this->Inputs()) {
-          if (!input_tensor->IsCPU()) {
-#ifdef TRITON_ENABLE_GPU
-            input_tensor->SetCudaIpcMutexes(
-                stub->CudaIpcOpenMutex(), stub->CudaIpcCloseMutex());
-            input_tensor->LoadGPUData(shm_pool);
-#endif  // TRITON_ENABLE_GPU
-          }
-        }
-      }
-      catch (const PythonBackendException& exception) {
-        // We need to catch the exception here. Otherwise, we will not notify
-        // the main process and it will wait for the resposne forever.
-        pb_exception = exception;
-        has_exception = true;
-      }
-
-      {
-        bi::scoped_lock<bi::interprocess_mutex> lock{
-            *(ipc_message->ResponseMutex())};
-        ipc_message->ResponseCondition()->notify_all();
-        ipc_message->ResponseCondition()->wait(lock);
-      }
-    }
-
-    // The exception will be thrown after the message was sent to the main
-    // process.
-    if (has_exception) {
-      throw pb_exception;
-    }
-
-    // Get the response for the current message.
-    std::unique_ptr<IPCMessage> bls_response = IPCMessage::LoadFromSharedMemory(
-        shm_pool, ipc_message->RequestOffset());
-    shm_pool->MapOffset((char**)&response_batch, bls_response->Args());
-    responses_is_set = true;
-
-    if (response_batch->has_error) {
-      if (response_batch->is_error_set) {
-        char* err_string;
-        LoadStringFromSharedMemory(shm_pool, response_batch->error, err_string);
-        return std::make_unique<InferResponse>(
-            std::vector<std::shared_ptr<PbTensor>>{},
-            std::make_shared<PbError>(err_string));
-      } else {
-        return std::make_unique<InferResponse>(
-            std::vector<std::shared_ptr<PbTensor>>{},
-            std::make_shared<PbError>(
-                "An error occurred while performing BLS request."));
-      }
-    }
-  }
-  catch (const PythonBackendException& pb_exception) {
-    return std::make_unique<InferResponse>(
-        std::vector<std::shared_ptr<PbTensor>>{},
-        std::make_shared<PbError>(pb_exception.what()));
+  std::vector<std::string> requested_output_names;
+  for (size_t output_idx = 0;
+       output_idx < infer_request_shm_ptr_->requested_output_count;
+       ++output_idx) {
+    auto& pb_string = requested_output_names_shm[output_idx];
+    requested_output_names.emplace_back(pb_string->String());
   }
 
-  if (responses_is_set) {
-    std::unique_ptr<InferResponse> infer_response =
-        InferResponse::LoadFromSharedMemory(
-            shm_pool, response_batch->responses, stub->CudaIpcOpenMutex(),
-            stub->CudaIpcCloseMutex());
-
-    return infer_response;
-  } else {
-    return std::make_unique<InferResponse>(
-        std::vector<std::shared_ptr<PbTensor>>{},
-        std::make_shared<PbError>(
-            "An error occurred while performing BLS request."));
-  }
+  request_id_ = request_id_shm_->String();
+  requested_output_names_ = std::move(requested_output_names);
+  model_name_ = model_name_shm_->String();
+  flags_ = infer_request_shm_ptr_->flags;
+  model_version_ = infer_request_shm_ptr_->model_version;
+  correlation_id_ = infer_request_shm_ptr_->correlation_id;
 }
 
-#endif
 }}}  // namespace triton::backend::python