--- a/src/infer_request.h
+++ b/src/infer_request.h
@@ -27,19 +27,29 @@
 #pragma once
 
 #include <string>
-#include "infer_response.h"
 #include "pb_tensor.h"
 
 namespace triton { namespace backend { namespace python {
-class InferRequest {
-  std::string request_id_;
-  uint64_t correlation_id_;
-  std::vector<std::shared_ptr<PbTensor>> inputs_;
-  std::vector<std::string> requested_output_names_;
-  std::string model_name_;
-  int64_t model_version_;
-  uint32_t flags_;
 
+//
+// Inference Request
+//
+struct InferRequestShm {
+  // Offset for the id field.
+  bi::managed_external_buffer::handle_t id;
+  uint64_t correlation_id;
+  // Offset for input field.
+  bi::managed_external_buffer::handle_t inputs;
+  uint32_t input_count;
+  // Offset for the requested output names
+  bi::managed_external_buffer::handle_t requested_output_names;
+  uint32_t requested_output_count;
+  bi::managed_external_buffer::handle_t model_name;
+  int64_t model_version;
+  uint32_t flags;
+};
+
+class InferRequest {
  public:
   InferRequest(
       const std::string& request_id, uint64_t correlation_id,
@@ -59,20 +69,47 @@ class InferRequest {
 
   /// Save an Inference Request to shared memory.
   /// \param shm_pool Shared memory pool to save the inference request.
-  /// \param request_shm A pointer to a location in shared memory with enough
-  /// space to save the inference request.
-  void SaveToSharedMemory(
-      std::unique_ptr<SharedMemory>& shm_pool, Request* request_shm);
+  void SaveToSharedMemory(std::unique_ptr<SharedMemoryManager>& shm_pool);
 
   /// Create an Inference Request object from shared memory.
   /// \param shm_pool Shared memory pool
   /// \param request_offset Shared memory offset of the request.
   static std::unique_ptr<InferRequest> LoadFromSharedMemory(
-      std::unique_ptr<SharedMemory>& shm_pool, off_t request_offset,
-      std::shared_ptr<std::mutex>& cuda_ipc_open_mutex,
-      std::shared_ptr<std::mutex>& cuda_ipc_close_mutex);
-#ifdef TRITON_PB_STUB
-  std::unique_ptr<InferResponse> Exec();
-#endif
+      std::unique_ptr<SharedMemoryManager>& shm_pool,
+      bi::managed_external_buffer::handle_t request_offset);
+
+ private:
+  InferRequest(
+      AllocatedSharedMemory<InferRequestShm>& infer_request_shm,
+      std::unique_ptr<PbString>& request_id_shm,
+      std::vector<std::unique_ptr<PbString>>& requested_output_names_shm,
+      std::unique_ptr<PbString>& model_name_shm,
+      AllocatedSharedMemory<bi::managed_external_buffer::handle_t>&
+          output_names_handle_shm,
+      AllocatedSharedMemory<bi::managed_external_buffer::handle_t>&
+          input_tensors_handle,
+      std::vector<std::shared_ptr<PbTensor>>& input_tensors);
+
+  std::string request_id_;
+  uint64_t correlation_id_;
+  std::vector<std::shared_ptr<PbTensor>> inputs_;
+  std::vector<std::string> requested_output_names_;
+  std::string model_name_;
+  int64_t model_version_;
+  uint32_t flags_;
+
+  // Shared Memory Data Structures
+  AllocatedSharedMemory<InferRequestShm> infer_request_shm_;
+  InferRequestShm* infer_request_shm_ptr_;
+
+  std::unique_ptr<PbString> request_id_shm_;
+  std::vector<std::unique_ptr<PbString>> requested_output_names_shm_;
+  std::unique_ptr<PbString> model_name_shm_;
+  AllocatedSharedMemory<bi::managed_external_buffer::handle_t>
+      output_names_handle_shm_;
+  bi::managed_external_buffer::handle_t* output_names_handle_shm_ptr_;
+  AllocatedSharedMemory<bi::managed_external_buffer::handle_t>
+      input_tensors_handle_;
+  bi::managed_external_buffer::handle_t* input_tensors_handle_ptr_;
 };
 }}};  // namespace triton::backend::python