Performance results before and after these changes:

|  Scenario  |  Latency  (before) |  Throughput (before)  | Latency  (after) |  Throughput (after)  |
|:--- |:--- | :--- |:--- | :--- |
| 16mb_latency_grpc | 39.714 | 25.2 |  40.229 | 25.0 |
| 16mb_latency_http | 41.374 |  24.2 |  47.424 | 21.0 |
| 16mb_throughput_grpc |  145.534 |  110.2 |  145.975 | 109.6 |
| 16mb_throughput_http |  404.136 |    38.4  |   291.94 |   52.8 |
| max_throughput_grpc | 1.481 |  10772.8 |  2.299 |  6948.6 ⬇️  |
| max_throughput_http |   1.02 |  15603.6 |   1.666 |  9573.0 ⬇️  |
| min_latency_grpc |   0.386 |  2562.0 |   0.481 |  2062.6  ⬇️ |
| min_latency_http | 0.443 |   2237.8 |   0.454 |   2185.8 ⬇️ |

It looks like there is a significant slowdown for the small tensor no model benchmarks but the performance for large payload seems to be on par with before.

## Update: The numbers below are collected using the C-API
|  Scenario  |  Latency  (before) |  Throughput (before)  | Latency  (after) |  Throughput (after)  |
|:--- |:--- | :--- |:--- | :--- |
| 16mb_latency | 11624 (us) | 86 infer/sec |  11684 usec |   85.6 infer/sec  |
| 16mb_throughput | 96059  (us) |  166.4 infer/sec  |  95579 (us)  |167.2 infer/sec |
| max_throughput |  972 (us) |  **16435.4 infer/sec**  |1078 (us)  |   **14824.4 infer/sec** |
| min_latency |   136 (us) |  **7323.4 infer/sec** | 163 (us) |  **6092.2 infer/sec** | 

Still there is a gap but the results are much closer now to the previous numbers. The PA number without the C-API for `max_throughput_http` is  for 12766.8 infer/sec throughput.  This is up from 9573.0 in the previous results.